+++ Actions/PhotonNetworkGetRoomListProperties.cs
@@ -11,7 +11,6 @@ namespace HutongGames.PlayMaker.Actions
 	public class PhotonNetworkGetRoomListProperties : FsmStateAction
 	{
 
-
 		[ActionSection("Builtin Properties")]
 		[Tooltip("All rooms' name")]
 		[UIHint(UIHint.Variable)]
@@ -21,7 +20,7 @@ namespace HutongGames.PlayMaker.Actions
 		[Tooltip("All rooms' playerCount")]
 		[UIHint(UIHint.Variable)]
 		[ArrayEditor(VariableType.Int)]
-		public FsmArray playerCount;
+		public FsmArray playerCounts;
 
 		[Tooltip("All rooms' maxPlayers")]
 		[UIHint(UIHint.Variable)]
@@ -31,8 +30,14 @@ namespace HutongGames.PlayMaker.Actions
 		[Tooltip("All rooms' open")]
 		[UIHint(UIHint.Variable)]
 		[ArrayEditor(VariableType.Bool)]
-		public FsmArray open;
+		public FsmArray opens;
 
+
+		[Tooltip("Custom Properties you have assigned to rooms.")]
+		[CompoundArray("Rooms Custom Properties", "property", "value")]
+		public FsmString[] customPropertyKeys;
+		[UIHint(UIHint.Variable)]
+		public FsmArray[] customPropertiesValues;
 
 		public FsmEvent notInLobbyEvent;
 
@@ -42,10 +47,16 @@ namespace HutongGames.PlayMaker.Actions
 		{
 
 			names = new FsmArray() {UseVariable=true};
-			playerCount = new FsmArray() {UseVariable=true};
+			playerCounts = new FsmArray() {UseVariable=true};
 			maxPlayers = new FsmArray() {UseVariable=true};
-			open = new FsmArray() {UseVariable=true};
+			opens = new FsmArray() {UseVariable=true};
+
+			customPropertyKeys = null;
+			customPropertiesValues = null;
+
+			notInLobbyEvent = null;
 		}
+
 		
 		public override void OnEnter()
 		{
@@ -60,8 +71,16 @@ namespace HutongGames.PlayMaker.Actions
 
 			rooms = PhotonNetwork.GetRoomList();
 
+			for(int x=0;x<customPropertiesValues.Length;x++)
+			{
+				if (! customPropertiesValues[x].IsNone) customPropertiesValues[x].Resize(rooms.Length);
+			}
+
+
 
 			if (!names.IsNone) names.Resize(rooms.Length);
+			if (!playerCounts.IsNone) playerCounts.Resize(rooms.Length);
+			if (!maxPlayers.IsNone) playerCounts.Resize(maxPlayers.Length);
 
 			int i=0;
 			
@@ -69,12 +88,32 @@ namespace HutongGames.PlayMaker.Actions
 			{
 
 				if (!names.IsNone) names.Set(i,room.name);
-			
+				if (!playerCounts.IsNone) playerCounts.Set(i,room.playerCount);
+				if (!maxPlayers.IsNone) maxPlayers.Set(i,room.maxPlayers);
+
+				// get the custom properties
+				int k = 0;
+				foreach(FsmString key in customPropertyKeys)
+				{
+					if (room.customProperties.ContainsKey(key.Value) && ! customPropertiesValues[k].IsNone)
+					{
+						customPropertiesValues[k].Set(i,room.customProperties[key.Value]);
+					}
+					k++;
+				}
+
+
 				i++;
 			}
 
 			if (!names.IsNone)  names.SaveChanges();
+			if (!playerCounts.IsNone)  playerCounts.SaveChanges();
+			if (!maxPlayers.IsNone)  maxPlayers.SaveChanges();
 
+			for(int x=0;x<customPropertiesValues.Length;x++)
+			{
+				if (! customPropertiesValues[x].IsNone) customPropertiesValues[x].SaveChanges();
+			}
 		}
 		
 	}